+++ mongodb/src/main/scala/quasar/physical/mongodb/execution.scala
@@ -21,6 +21,7 @@ import quasar.common.SortDir
 import quasar.fp._
 import quasar.physical.mongodb.accumulator._
 import quasar.physical.mongodb.expression._
+import quasar.physical.mongodb.selector.Selector
 import quasar.physical.mongodb.workflow._
 
 import scalaz._, Scalaz._
