+++ mongodb/src/main/scala/quasar/physical/mongodb/mapreduce.scala
@@ -19,6 +19,7 @@ package quasar.physical.mongodb
 import slamdata.Predef._
 import quasar.common.SortDir
 import quasar.javascript._
+import quasar.physical.mongodb.selector.Selector
 
 import com.mongodb.client.model.MapReduceAction
 import monocle.macros.GenLens
