+++ mongodb/src/main/scala/quasar/physical/mongodb/optimize/optimize.scala
@@ -22,6 +22,7 @@ import quasar.fp._
 import quasar.fp.ski._
 import quasar.physical.mongodb.accumulator._
 import quasar.physical.mongodb.expression._
+import quasar.physical.mongodb.selector.Selector
 import quasar.physical.mongodb.workflow._
 
 import matryoshka._
