+++ mongodb/src/main/scala/quasar/physical/mongodb/workflowbuilder.scala
@@ -28,6 +28,7 @@ import quasar.javascript._
 import quasar.jscore, jscore.JsFn
 import quasar.physical.mongodb.accumulator._
 import quasar.physical.mongodb.expression._
+import quasar.physical.mongodb.selector.Selector
 import quasar.physical.mongodb.workflow._
 import quasar.qscript.{IdStatus, ShiftType}
 import quasar.std.StdLib._
